Update plugin framework

+++ webpack.config.js
@@ -13,8 +13,15 @@ function readManifest(manifestPath) {
 }
 
 function createPluginArchive(sourceDir, destPath) {
-	const distFiles = glob.sync(`${sourceDir}/**/*`)
+	const distFiles = glob.sync(`${sourceDir}/**/*`, { nodir: true })
 		.map(f => f.substr(sourceDir.length + 1));
+
+	if (!distFiles.length) {
+		// Usually means there's an error, which is going to be printed by
+		// webpack
+		console.info('Plugin archive was not created because the "dist" directory is empty');
+		return;
+	}
 
 	fs.removeSync(destPath);
 
@@ -32,13 +39,16 @@ function createPluginArchive(sourceDir, destPath) {
 	console.info(`Plugin archive has been created in ${destPath}`);
 }
 
-const distDir = path.resolve(__dirname, 'dist');
-const srcDir = path.resolve(__dirname, 'src');
+const rootDir = path.resolve(__dirname);
+const distDir = path.resolve(rootDir, 'dist');
+const srcDir = path.resolve(rootDir, 'src');
 const manifestPath = `${srcDir}/manifest.json`;
 const manifest = readManifest(manifestPath);
 const archiveFilePath = path.resolve(__dirname, `${manifest.id}.jpl`);
 
-var config = {
+fs.removeSync(distDir);
+
+const baseConfig = {
 	mode: 'production',
 	target: 'node',
 	module: {
@@ -50,21 +60,23 @@ var config = {
 			},
 		],
 	},
+};
+
+const pluginConfig = Object.assign({}, baseConfig, {
+	entry: './src/index.ts',
 	resolve: {
 		alias: {
 			api: path.resolve(__dirname, 'api'),
 		},
 		extensions: ['.tsx', '.ts', '.js'],
 	},
-};
-
-var index = Object.assign({}, config, {
-	name: 'index',
-	entry: './src/index.ts',
 	output: {
 		filename: 'index.js',
 		path: distDir,
 	},
+});
+
+const lastStepConfig = {
 	plugins: [
 		new CopyPlugin({
 			patterns: [
@@ -74,31 +86,83 @@ var index = Object.assign({}, config, {
 					to: path.resolve(__dirname, 'dist'),
 					globOptions: {
 						ignore: [
+							// All TypeScript files are compiled to JS and
+							// already copied into /dist so we don't copy them.
 							'**/*.ts',
 							'**/*.tsx',
+
+							// Currently we don't support JS files for the main
+							// plugin script. We support it for content scripts,
+							// but theyr should be declared in manifest.json,
+							// and then they are also compiled and copied to
+							// /dist. So wse also don't need to copy JS files.
+							'**/*.js',
 						],
 					},
 				},
 			],
 		}),
-	],
-});
-
-var contentScript = Object.assign({}, config, {
-	name: 'contentScript',
-	entry: './src/mathMode.ts',
-	output: {
-		filename: 'mathMode.js',
-		path: distDir,
-		library: 'default',
-		libraryTarget: 'commonjs',
-		libraryExport: 'default',
-	},
-	plugins: [
 		new WebpackOnBuildPlugin(function() {
 			createPluginArchive(distDir, archiveFilePath);
 		}),
 	],
+};
+
+const contentScriptConfig = Object.assign({}, baseConfig, {
+	resolve: {
+		alias: {
+			api: path.resolve(__dirname, 'api'),
+		},
+		extensions: ['.tsx', '.ts', '.js'],
+	},
 });
 
-module.exports = [index, contentScript];
+function resolveContentScriptPaths(name) {
+	if (['.js', '.ts', '.tsx'].includes(path.extname(name).toLowerCase())) {
+		throw new Error(`Content script path must not include file extension: ${name}`);
+	}
+
+	const pathsToTry = [
+		`./src/${name}.ts`,
+		`${'./src/' + '/'}${name}.js`,
+	];
+
+	for (const pathToTry of pathsToTry) {
+		if (fs.pathExistsSync(`${rootDir}/${pathToTry}`)) {
+			return {
+				entry: pathToTry,
+				output: {
+					filename: `${name}.js`,
+					path: distDir,
+					library: 'default',
+					libraryTarget: 'commonjs',
+					libraryExport: 'default',
+				},
+			};
+		}
+	}
+
+	throw new Error(`Could not find content script "${name}" at locations ${JSON.stringify(pathsToTry)}`);
+}
+
+function createContentScriptConfigs() {
+	if (!manifest.content_scripts) return [];
+
+	const output = [];
+
+	for (const contentScriptName of manifest.content_scripts) {
+		const scriptPaths = resolveContentScriptPaths(contentScriptName);
+		output.push(Object.assign({}, contentScriptConfig, {
+			entry: scriptPaths.entry,
+			output: scriptPaths.output,
+		}));
+	}
+
+	return output;
+}
+
+const exportedConfigs = [pluginConfig].concat(createContentScriptConfigs());
+
+exportedConfigs[exportedConfigs.length - 1] = Object.assign({}, exportedConfigs[exportedConfigs.length - 1], lastStepConfig);
+
+module.exports = exportedConfigs;
